+++ react/modules/schema.ts
@@ -59,6 +59,12 @@ const IMAGE_LIST_MESSAGES = defineMessages({
   analyticsPromotionPosition: {
     id: 'admin/editor.image.analytics.promotionPosition',
   },
+  analyticsProductId: {
+    id: 'admin/editor.image.analytics.productId',
+  },
+  analyticsProductName: {
+    id: 'admin/editor.image.analytics.productName',
+  },
 })
 
 export const IMAGE_LIST_SCHEMA = {